* origin/2.7: (39 commits)
  UObject: refuse non-String keyed Dictionaries.
  dev: use and fix diff.
  build: please picky compilers.
  Dev: fix colordiff test.
  dev: avoid colors where we can't have them.
  modules: update.
  tests: update.
  ast: fix ignores.
  Java: work around a Boost 1.48 bug.
  clang: be compatible.
  libport: update.
  Compiler: fix warning.
  build: please clang.
  build: please clang++: use this-> where appropriate.
  Java: style changes.
  UMessage: more consistent error messages.
  Build: don't hurt GCC's feelings with clang.
  Build: please clang.
  build: please clang++.
  Style changes.
  ...

Conflicts:
	dev/tools.py
	doc/guide/k1-to-k2.tex
	doc/tables/release-notes.tex
	include/kernel/fwd.hh
	sdk-remote/libport
	sdk-remote/src/swig/java/src/call-macros.hh
	sdk-remote/src/swig/java/src/callbacks-caller.cc
	sdk-remote/src/swig/java/src/callbacks-caller.hh
	sdk-remote/src/swig/java/src/converter.hh
	sdk-remote/src/swig/java/swig/java.i
	sdk-remote/src/swig/java/tests/all/All.java
	src/object/finalizable.cc
	src/object/path.cc
	src/runner/interpreter-apply.cc
	src/runner/interpreter-visit.hxx
	submodules/modules
	tests

+++ sdk-remote/src/swig/java/tests/all/All.java
@@ -580,7 +580,7 @@ public class All extends UObject
     /** Test function parameter and return value **/
     public double transmitD(double v)
     {
-	return -(double)v;
+	return -v;
     }
 
     public UList transmitL(UList l)
